NLP refactoring patches.
