+++ examples/nlp/intent_detection_slot_tagging/joint_intent_slot_infer.py
@@ -34,9 +34,9 @@ parser.add_argument("--local_rank", default=None, type=int)
 parser.add_argument("--batch_size", default=128, type=int)
 parser.add_argument("--max_seq_length", default=50, type=int)
 parser.add_argument("--pretrained_bert_model", default="bert-base-uncased", type=str)
-parser.add_argument("--dataset_name", default='snips-all', type=str)
-parser.add_argument("--data_dir", default='data/nlu/snips', type=str)
-parser.add_argument("--work_dir", required=True, help="your checkpoint folder", type=str)
+parser.add_argument("--dataset_name", default='atis', type=str)
+parser.add_argument("--data_dir", default='data/atis', type=str)
+parser.add_argument("--checkpoint_dir", required=True, help="your checkpoint folder", type=str)
 parser.add_argument("--eval_file_prefix", default='test', type=str)
 parser.add_argument("--amp_opt_level", default="O0", type=str, choices=["O0", "O1", "O2"])
 parser.add_argument("--do_lower_case", action='store_false')
@@ -58,7 +58,7 @@ pretrained_bert_model = BERT(pretrained_model_name=args.pretrained_bert_model)
 hidden_size = pretrained_bert_model.hidden_size
 tokenizer = BertTokenizer.from_pretrained(args.pretrained_bert_model)
 
-data_desc = JointIntentSlotDataDesc(args.data_dir, args.do_lower_case, args.dataset_name)
+data_desc = JointIntentSlotDataDesc(data_dir=args.data_dir)
 
 # Evaluation pipeline
 logging.info("Loading eval data...")
@@ -96,7 +96,7 @@ evaluated_tensors = nf.infer(
         input_data.intents,
         input_data.slots,
     ],
-    checkpoint_dir=args.work_dir,
+    checkpoint_dir=args.checkpoint_dir,
 )
